Added online progress save.
When you upgrade your car, its data is sent to an api which saves it to the file. When you then need this data, for example for a race, get requests will get you that. So now you can login from anywhere(If the server is online) and get your car and coins. Features done 2/3